Fix function argument type

+++ database/functions/create-payroll-cover.sql
@@ -1,7 +1,7 @@
 DROP FUNCTION IF EXISTS create_payroll_cover;
 CREATE FUNCTION create_payroll_cover (
     employee_id int,
-    employee_hourly_rate int,
+    employee_hourly_rate float,
     registered_hours float,
     deliveries_count int,
     payment_date timestamptz
